Switching Int Ecu ==> Float Ecu for User and all others classes and test classes.

+++ backend/tinyfarm/src/main/java/com/api/tinyfarm/controller/UserController.java
@@ -2,15 +2,15 @@ package com.api.tinyfarm.controller;
 
 import com.api.tinyfarm.model.User;
 import com.api.tinyfarm.service.UserService;
+import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @RestController
 @RequestMapping("/api/users")
 public class UserController {
+
     private final UserService userService;
 
     public UserController(UserService userService) {
@@ -19,10 +19,9 @@ public class UserController {
 
     @GetMapping("")
     public ResponseEntity<List<User>> getAll() {
-
         try {
             return ResponseEntity.ok(userService.findAll());
-        }catch(Exception e){
+        } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
     }
@@ -31,29 +30,32 @@ public class UserController {
     public ResponseEntity<User> getById(@PathVariable Long id) {
         try {
             return ResponseEntity.ok(userService.findById(id));
-        }catch(Exception e){
+        } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
     }
 
     @PostMapping("")
     public ResponseEntity<User> create(@RequestBody User user) {
-
         try {
             return ResponseEntity.ok(userService.create(user));
-        }catch (IllegalArgumentException e){
+        } catch (IllegalArgumentException e) {
             return ResponseEntity.status(HttpStatus.CONFLICT).build();
-        }catch (Exception e){
-            return  ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        } catch (Exception e) {
+            return ResponseEntity.status(
+                HttpStatus.INTERNAL_SERVER_ERROR
+            ).build();
         }
     }
 
     @PutMapping("/id/{id}")
-    public ResponseEntity<User> update(@PathVariable Long id,
-                                       @RequestBody User user) {
+    public ResponseEntity<User> update(
+        @PathVariable Long id,
+        @RequestBody User user
+    ) {
         try {
             return ResponseEntity.ok(userService.update(id, user));
-        }catch(Exception e){
+        } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
     }
@@ -64,29 +66,33 @@ public class UserController {
         try {
             userService.delete(id);
             return ResponseEntity.noContent().build();
-        }catch(Exception e){
+        } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
     }
 
     // PATCH /users/1/ecus?amount=8 → ajouter des écus (vente d'un oeuf)
     @PatchMapping("/ecus/add/id/{id}")
-    public ResponseEntity<User> addEcus(@PathVariable Long id,
-                                             @RequestParam Integer amount) {
+    public ResponseEntity<User> addEcus(
+        @PathVariable Long id,
+        @RequestParam Float amount
+    ) {
         try {
             return ResponseEntity.ok(userService.addEcus(id, amount));
-        }catch(Exception e){
+        } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
     }
 
     // PATCH /users/1/ecus/retirer?amount=3 → retirer des écus (nourrir un animal)
     @PatchMapping("/ecus/withdraw/id/{id}")
-    public ResponseEntity<User> withdrawEcus(@PathVariable Long id,
-                                             @RequestParam Integer amount) {
+    public ResponseEntity<User> withdrawEcus(
+        @PathVariable Long id,
+        @RequestParam Float amount
+    ) {
         try {
             return ResponseEntity.ok(userService.withdrawEcus(id, amount));
-        }catch(Exception e){
+        } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
     }
